refactor(#311): RegularApplyList 컴포넌트 분리
